Created CRUD methods for User Added models for the 6 other classes (7 total) Commented all helpful tools and annotations to streamline future progress

+++ src/main/java/com/masongarrett/taskmanagementsystem/dao/UserDAO.java
@@ -7,9 +7,9 @@ import com.masongarrett.taskmanagementsystem.model.User;
 public interface UserDAO {
 
     List<User> get();
-    User get(int id);
+    User get(long id);
     void save(User user);
-    void delete(int id);
+    void delete(long id);
 
 
 }
